Move mock model into importable package
