+++ src/pykeen/models/__init__.py
@@ -10,6 +10,7 @@ from class_resolver import Resolver, get_subclasses
 
 from .base import EntityRelationEmbeddingModel, Model, _OldAbstractModel
 from .baseline import EvaluationOnlyModel, MarginalDistributionBaseline
+from .mocks import MockModel
 from .multimodal import ComplExLiteral, DistMultLiteral, DistMultLiteralGated, LiteralModel
 from .nbase import ERModel, _NewAbstractModel
 from .resolve import make_model, make_model_cls
@@ -74,6 +75,7 @@ __all__ = [
     "ERMLPE",
     "HolE",
     "KG2E",
+    "MockModel",
     "MuRE",
     "NodePiece",
     "NTN",
