+++ tests/mocks.py
@@ -9,14 +9,11 @@ from torch import nn
 
 from pykeen.evaluation import Evaluator, MetricResults, RankBasedMetricResults
 from pykeen.evaluation.rank_based_evaluator import RANK_REALISTIC, RANK_TYPES, SIDES
-from pykeen.models import EntityRelationEmbeddingModel, Model
-from pykeen.nn.emb import EmbeddingSpecification, RepresentationModule
-from pykeen.triples import CoreTriplesFactory
+from pykeen.nn.emb import RepresentationModule
 from pykeen.typing import MappedTriples
 
 __all__ = [
     "CustomRepresentations",
-    "MockModel",
 ]
 
 
@@ -30,39 +27,6 @@ class CustomRepresentations(RepresentationModule):
     def forward(self, indices: Optional[torch.LongTensor] = None) -> torch.FloatTensor:  # noqa:D102
         n = self.max_id if indices is None else indices.shape[0]
         return self.x.unsqueeze(dim=0).repeat(n, *(1 for _ in self.shape))
-
-
-class MockModel(EntityRelationEmbeddingModel):
-    """A mock model returning fake scores."""
-
-    def __init__(self, *, triples_factory: CoreTriplesFactory):
-        super().__init__(
-            triples_factory=triples_factory,
-            entity_representations=EmbeddingSpecification(embedding_dim=50),
-            relation_representations=EmbeddingSpecification(embedding_dim=50),
-        )
-        num_entities = self.num_entities
-        self.scores = torch.arange(num_entities, dtype=torch.float, requires_grad=True)
-        self.num_backward_propagations = 0
-
-    def _generate_fake_scores(self, batch: torch.LongTensor) -> torch.FloatTensor:
-        """Generate fake scores s[b, i] = i of size (batch_size, num_entities)."""
-        batch_size = batch.shape[0]
-        batch_scores = self.scores.view(1, -1).repeat(batch_size, 1)
-        assert batch_scores.shape == (batch_size, self.num_entities)
-        return batch_scores
-
-    def score_hrt(self, hrt_batch: torch.LongTensor) -> torch.FloatTensor:  # noqa: D102
-        return self.scores[torch.randint(high=self.num_entities, size=hrt_batch.shape[:-1])]
-
-    def score_t(self, hr_batch: torch.LongTensor) -> torch.FloatTensor:  # noqa: D102
-        return self._generate_fake_scores(batch=hr_batch)
-
-    def score_h(self, rt_batch: torch.LongTensor) -> torch.FloatTensor:  # noqa: D102
-        return self._generate_fake_scores(batch=rt_batch)
-
-    def reset_parameters_(self) -> Model:  # noqa: D102
-        pass  # Not needed for unittest
 
 
 class MockEvaluator(Evaluator):
